Affichage d'un utilisateur et classe Aop cree

+++ src/main/java/com/user/userapi/dao/UserRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.user.userapi.model.Utilisateur;
 
-public interface UserRepository extends JpaRepository<Utilisateur, Long> {
+public interface UserRepository extends JpaRepository<Utilisateur, Integer> {
+
+	
 
 }
